@@ -4,7 +4,7 @@ homepage: https://github.com/qiniu/dart-sdk
 description: The sdk basic of Qiniu products
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.11.0-213.0.dev <3.0.0'
 
 dependencies:
   dio: ^3.0.10
@@ -13,5 +13,5 @@ dependencies:
 
 dev_dependencies:
   pedantic: ^1.9.0
-  test: ^1.14.4
   dotenv: ^1.0.0
+  test: ^1.14.4
