@@ -7,7 +7,7 @@ export 'put_policy.dart';
 
 class TokenInfo {
   final String accessKey;
-  final PutPolicy putPolicy;
+  final PutPolicy? putPolicy;
   const TokenInfo(this.accessKey, this.putPolicy);
 }
 
@@ -27,19 +27,16 @@ class Auth {
   final String secretKey;
 
   const Auth({
-    @required this.accessKey,
-    @required this.secretKey,
-  })  : assert(accessKey != null),
-        assert(secretKey != null);
+    required this.accessKey,
+    required this.secretKey,
+  });
 
   /// 根据上传策略生成上传使用的 Token。
   ///
   /// 具体的上传策略说明请参考 [PutPolicy] 模块
   String generateUploadToken({
-    @required PutPolicy putPolicy,
+    required PutPolicy putPolicy,
   }) {
-    assert(putPolicy != null);
-
     var data = jsonEncode(putPolicy);
     var encodedPutPolicy = base64Url.encode(utf8.encode(data));
     var baseToken = generateAccessToken(bytes: utf8.encode(encodedPutPolicy));
@@ -52,14 +49,10 @@ class Auth {
   /// [deadline] 有效时间，单位为秒，例如 1451491200
   /// [bucketDomain] 空间绑定的域名，例如 http://test.bucket.com
   String generateDownloadToken({
-    @required String key,
-    @required int deadline,
-    @required String bucketDomain,
+    required String key,
+    required int deadline,
+    required String bucketDomain,
   }) {
-    assert(key != null);
-    assert(deadline != null);
-    assert(bucketDomain != null);
-
     var downloadURL = '$bucketDomain/$key?e=$deadline';
     return generateAccessToken(bytes: utf8.encode(downloadURL));
   }
@@ -67,8 +60,7 @@ class Auth {
   /// 根据数据签名，生成 Token（用于接口的访问鉴权）。
   ///
   /// 访问七牛的接口需要对请求进行签名, 该方法提供 Token 签发服务
-  String generateAccessToken({@required List<int> bytes}) {
-    assert(bytes != null);
+  String generateAccessToken({required List<int> bytes}) {
     var hmacEncoder = Hmac(sha1, utf8.encode(secretKey));
 
     var sign = hmacEncoder.convert(bytes);
@@ -80,13 +72,12 @@ class Auth {
   ///
   /// 从 Token 字符串中解析 [accessKey]、[PutPolicy] 信息
   static TokenInfo parseToken(String token) {
-    assert(token != null && token != '');
     var segments = token.split(':');
     if (segments.length < 2) {
       throw ArgumentError('invalid token');
     }
 
-    PutPolicy putPolicy;
+    PutPolicy? putPolicy;
     var accessKey = segments.first;
 
     /// 具体的 token 信息可以参考这里。
