@@ -2,10 +2,13 @@ part of 'put_parts_task.dart';
 
 /// initParts 的返回体
 class InitParts {
-  String uploadId;
-  int expireAt;
+  final int expireAt;
+  final String uploadId;
 
-  InitParts({this.uploadId, this.expireAt});
+  InitParts({
+    required this.expireAt,
+    required this.uploadId,
+  });
 
   factory InitParts.fromJson(Map json) {
     return InitParts(
@@ -15,51 +18,63 @@ class InitParts {
   }
 
   Map<String, dynamic> toJson() {
-    return {'uploadId': uploadId, 'expireAt': expireAt};
+    return <String, dynamic>{
+      'uploadId': uploadId,
+      'expireAt': expireAt,
+    };
   }
 }
 
 /// 初始化一个分片上传任务，为 [UploadPartsTask] 提供 uploadId
 class InitPartsTask extends RequestTask<InitParts> with CacheMixin<InitParts> {
-  String token;
-  String bucket;
-  String host;
-  String key;
-  File file;
+  final File file;
+  final String token;
+  final String bucket;
+  final String host;
+  final String? key;
 
   @override
-  String _cacheKey;
+  late final String _cacheKey;
 
   InitPartsTask({
-    this.host,
-    this.bucket,
-    this.token,
+    required this.file,
+    required this.host,
+    required this.bucket,
+    required this.token,
     this.key,
-    this.file,
   });
 
-  static String getCacheKey(String path, String key, int length) {
+  static String getCacheKey(String path, int length, String? key) {
     return 'qiniu_dart_sdk_init_parts_task_${path}_key_${key}_size_$length';
   }
 
   @override
   void preStart() {
-    _cacheKey = InitPartsTask.getCacheKey(file.path, key, file.lengthSync());
     super.preStart();
   }
 
   @override
   Future<InitParts> createTask() async {
-    final initParts = getCache();
-    if (initParts != null) {
-      return InitParts.fromJson(json.decode(initParts) as Map);
+    _cacheKey = InitPartsTask.getCacheKey(file.path, await file.length(), key);
+
+    final headers = {'Authorization': 'UpToken $token'};
+    final paramMap = <String, String>{'buckets': bucket};
+
+    final initPartsCache = getCache();
+    if (initPartsCache != null) {
+      return InitParts.fromJson(json.decode(initPartsCache) as Map);
     }
+
+    if (key != null) {
+      paramMap.addAll({'objects': base64Url.encode(utf8.encode(key!))});
+    }
+
     final response = await client.post<Map>(
-      '$host/buckets/$bucket/objects/${base64Url.encode(utf8.encode(key))}/uploads',
+      '$host/${paramMap.entries.join('/')}/uploads',
 
       /// 这里 data 不传，dio 不会触发 cancel 事件
-      data: {},
-      options: Options(headers: {'Authorization': 'UpToken $token'}),
+      data: <String, dynamic>{},
+      options: Options(headers: headers),
     );
 
     return InitParts.fromJson(response.data);
