@@ -1,45 +1,50 @@
 import 'dart:io';
-
 import 'package:dio/dio.dart';
-import 'package:meta/meta.dart';
-
 import 'request_task.dart';
 
+// 直传请求响应
 class Put {
-  String key;
-  String hash;
+  final String key;
+  final String hash;
 
-  Put({this.key, this.hash});
+  Put({
+    required this.key,
+    required this.hash,
+  });
 
   factory Put.fromJson(Map json) {
-    return Put(key: json['key'] as String, hash: json['hash'] as String);
+    return Put(
+      key: json['key'] as String,
+      hash: json['hash'] as String,
+    );
   }
 }
 
+// 直传任务
 class PutTask extends RequestTask<Put> {
-  /// 上传凭证
-  String token;
-
-  // 资源名。如果不传则后端自动生成
-  String key;
-
   /// 上传文件
-  File file;
+  final File file;
+
+  /// 上传凭证
+  final String token;
+
+  // 资源名, 如果不传则后端自动生成
+  final String? key;
 
   PutTask({
-    @required this.token,
-    @required this.file,
+    required this.file,
+    required this.token,
     this.key,
-  })  : assert(token != null),
-        assert(file != null);
+  });
 
   @override
   Future<Put> createTask() async {
-    final formData = FormData.fromMap({
+    final formData = FormData.fromMap(<String, dynamic>{
+      'file': await MultipartFile.fromFile(file.path),
       'token': token,
       'key': key,
-      'file': await MultipartFile.fromFile(file.path)
     });
+
     final host = await config.hostProvider.getUpHost(token: token);
     final response = await client.post<Map>(host, data: formData);
 
