@@ -7,17 +7,15 @@ part 'host.dart';
 part 'cache.dart';
 
 class Config {
-  HostProvider hostProvider;
-  CacheProvider cacheProvider;
-  HttpClientAdapter httpClientAdapter;
+  final HostProvider hostProvider;
+  final CacheProvider cacheProvider;
+  final HttpClientAdapter httpClientAdapter;
 
   Config({
-    this.hostProvider,
-    this.cacheProvider,
-    this.httpClientAdapter,
-  }) {
-    hostProvider = hostProvider ?? DefaultHostProvider();
-    cacheProvider = cacheProvider ?? DefaultCacheProvider();
-    httpClientAdapter = httpClientAdapter ?? DefaultHttpClientAdapter();
-  }
+    HostProvider? hostProvider,
+    CacheProvider? cacheProvider,
+    HttpClientAdapter? httpClientAdapter,
+  })  : hostProvider = hostProvider ?? DefaultHostProvider(),
+        cacheProvider = cacheProvider ?? DefaultCacheProvider(),
+        httpClientAdapter = httpClientAdapter ?? DefaultHttpClientAdapter();
 }
