@@ -2,10 +2,13 @@ part of 'put_parts_task.dart';
 
 /// 切片信息
 class Part {
-  String etag;
-  int partNumber;
+  final String etag;
+  final int partNumber;
 
-  Part({this.etag, this.partNumber});
+  Part({
+    required this.etag,
+    required this.partNumber,
+  });
 
   factory Part.fromJson(Map json) {
     return Part(
@@ -15,6 +18,9 @@ class Part {
   }
 
   Map<String, dynamic> toJson() {
-    return {'etag': etag, 'partNumber': partNumber};
+    return <String, dynamic>{
+      'etag': etag,
+      'partNumber': partNumber,
+    };
   }
 }
