@@ -12,10 +12,11 @@ import 'config.dart';
 void main() {
   configEnv();
 
-  Storage storage;
-  setUpAll(() {
-    storage = Storage();
-  });
+  final storage = Storage();
+
+  // setUpAll(() {
+  //   storage = Storage();
+  // });
 
   test('put should works well.', () async {
     final putTask = storage.putFile(
@@ -38,9 +39,9 @@ void main() {
       Future.delayed(Duration(milliseconds: 1), putTask.cancel);
       final response = await putTask.future;
       expect(response.key, 'test_for_put.txt');
-    } catch (err) {
-      expect(err, isA<DioError>());
-      expect(err.type, DioErrorType.CANCEL);
+    } catch (error) {
+      expect(error, isA<DioError>());
+      expect((error as DioError).type, DioErrorType.CANCEL);
     }
   }, skip: !isSensitiveDataDefined);
 
@@ -50,14 +51,16 @@ void main() {
       token,
       options: PutOptions(key: 'test_for_put.txt'),
     );
-    int _sent, _total;
+
+    int? _sent, _total;
     putTask.addProgressListener((sent, total) {
       _sent = sent;
       _total = total;
     });
+
     final response = await putTask.future;
     expect(response.key, 'test_for_put.txt');
-    expect(_sent / _total, equals(1));
+    expect(_sent! / _total!, equals(1));
   }, skip: !isSensitiveDataDefined);
 
   test('putParts should works well.', () async {
@@ -77,7 +80,7 @@ void main() {
     final putPartsTask = storage.putFileParts(
       File('test_resource/test_for_put_parts.mp4'),
       token,
-      options: PutPartsOptions(key: 'test_for_put_parts.mp4', partSize: 1),
+      options: PutPartsOptions(key: 'test_for_put_parts.mp4'),
     );
 
     final response = await putPartsTask.future;
@@ -97,8 +100,8 @@ void main() {
     Future.delayed(Duration(milliseconds: 1), putPartsTask.cancel);
     try {
       await putPartsTask.future;
-    } catch (err) {
-      expect(err.type, DioErrorType.CANCEL);
+    } catch (error) {
+      expect((error as DioError).type, DioErrorType.CANCEL);
     }
     expect(putPartsTask.future, throwsA(TypeMatcher<DioError>()));
   }, skip: !isSensitiveDataDefined);
@@ -115,9 +118,9 @@ void main() {
 
     try {
       await putPartsTask.future;
-    } catch (err) {
-      expect(err, isA<DioError>());
-      expect(err.type, DioErrorType.CANCEL);
+    } catch (error) {
+      expect(error, isA<DioError>());
+      expect((error as DioError).type, DioErrorType.CANCEL);
     }
 
     expect(putPartsTask.future, throwsA(TypeMatcher<DioError>()));
@@ -146,7 +149,7 @@ void main() {
       host: await config.hostProvider.getUpHost(token: token),
 
       /// TOKEN_SCOPE 暂时只保存了 bucket 信息
-      bucket: env['QINIU_DART_SDK_TOKEN_SCOPE'],
+      bucket: env['QINIU_DART_SDK_TOKEN_SCOPE']!,
       key: key,
       file: file,
     );
@@ -167,16 +170,19 @@ void main() {
     expect(cacheProvider.value.length, 1);
 
     /// 初始化的缓存 key 生成逻辑
-    final cacheKey =
-        InitPartsTask.getCacheKey(file.path, key, file.lengthSync());
+    final cacheKey = InitPartsTask.getCacheKey(
+      file.path,
+      file.lengthSync(),
+      key,
+    );
 
     expect(cacheProvider.getItem(cacheKey), isA<String>());
 
     try {
       await putPartsTask.future;
-    } catch (err) {
-      expect(err, isA<DioError>());
-      expect(err.type, DioErrorType.CANCEL);
+    } catch (error) {
+      expect(error, isA<DioError>());
+      expect((error as DioError).type, DioErrorType.CANCEL);
     }
 
     final response = await storage
@@ -199,14 +205,14 @@ void main() {
       token,
       options: PutPartsOptions(key: 'test_for_put_parts.mp4', partSize: 1),
     );
-    int _sent, _total;
+    int? _sent, _total;
     putPartsTask.addProgressListener((sent, total) {
       _sent = sent;
       _total = total;
     });
     final response = await putPartsTask.future;
     expect(response, isA<CompleteParts>());
-    expect(_sent / _total, equals(1));
+    expect(_sent! / _total!, equals(1));
   }, skip: !isSensitiveDataDefined);
 }
 
@@ -236,7 +242,7 @@ class HttpAdapterTest extends HttpClientAdapter {
 
 class HostProviderTest extends HostProvider {
   @override
-  Future<String> getUpHost({String token}) async {
+  Future<String> getUpHost({required String token}) async {
     return 'https://upload-z2.qiniup.com';
   }
 }
